StandaloneMmPkg: add support to populate StMM boot data from device tree
Introduce support to populate StMM boot data via DTS parsing. The DTB is
passed as a boot argument by a binary of higer exception level.
Previously it was achieved by placing the boot data structure in a
shared buffer and the address of this shared buffer was passed by the
binary of higher exception level. Now either of the option can be used
for populating StMM boot info.

StMM boot information structure binding in device tree can be of following
prototype. Property values are not mentioned here.

bootarg {
  compatible = "bootargs";
  h_type  = <..>;
  h_version = <..>;
  h_size    = <..>;
  h_attr    = <..>;
  sp_mem_base         = <..>;
  sp_mem_limit        = <..>;
  sp_image_base       = <..>;
  sp_stack_base       = <..>;
  sp_heap_base        = <..>;
  sp_ns_comm_buf_base = <..>;
  sp_shared_buf_base  = <..>;
  sp_image_size       = <..>;
  sp_pcpu_stack_size  = <..>;
  sp_heap_size        = <..>;
  sp_ns_comm_buf_size = <..>;
  sp_shared_buf_size  = <..>;
  num_sp_mem_regions  = <..>;
  num_cpus            = <..>;
};

Addition of DTS supoort involves a dependency on FdtLib from EmbeddedPkg.

Signed-off-by: Sayanta Pattanayak <sayanta.pattanayak@arm.com>